+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Merged branches and closed issues
 
+- Measurement Control - Improve progress bar for iterative and batch run with tqdm. (!477, #346)
 - Deprecation - Explicit handling of np.arrays to avoid deprecated functionality when converting a dimensional array into scalar. (!476)
 - Documentation - Improve documentation build time and enable `sphinx-autobuild` (!471)
 - Documentation - Fix missing images in Jupyter cell outputs in documentation deployed using Gitlab Pages. (!480)